IMR-133 refactor : architectures of backend project
