+++ back/src/api/__init__.py
@@ -1,0 +1,10 @@
+from fastapi import APIRouter
+
+from src.api import feedback, music
+
+router = APIRouter()
+router.include_router(music.router, tags=["music"])
+router.include_router(feedback.router, tags=["feedback"])
+
+
+__all__ = ["router"]
