+++ back/src/dto/music.py
@@ -1,4 +1,5 @@
 from pydantic import BaseModel
+from fastapi import Form
 
 
 class RecommendMusic(BaseModel):
@@ -13,3 +14,11 @@ class RecommendMusic(BaseModel):
 class RecommendMusicResponse(BaseModel):
     session_id: str
     songs: list[RecommendMusic]
+
+
+class RecommendMusicRequest(BaseModel):
+    genres: list[str]
+
+    @classmethod
+    def as_form(cls, genres: list[str] = Form(...)):
+        return cls(genres=genres)
