+++ back/src/infer/song.py
@@ -1,12 +1,9 @@
 import os
 import re
-import ast
 import pandas as pd
-from collections import Counter
 from typing import List
 from huggingface_hub import Repository
-
-from src.utils import str2list
+from src.infer.utils import str2list
 
 
 class SongIdExtractor:
