+++ back/src/infer/spotify.py
@@ -2,7 +2,7 @@ import os
 import spotipy
 import pandas as pd
 from spotipy.oauth2 import SpotifyClientCredentials
-from src.utils import check_substring
+from src.infer.utils import check_substring
 from src.log.logger import get_spotify_logger
 
 logger = get_spotify_logger()
