+++ back/src/services/feedback.py
@@ -1,0 +1,17 @@
+from src.dto.feedback import UserFeedbackRequest
+from src.log.logger import get_feedback_logger
+
+
+class FeedbackService:
+    def __init__(self) -> None:
+        self.feedback_logger = get_feedback_logger()
+
+    def log_user_feedback(self, data: UserFeedbackRequest) -> None:
+        self.feedback_logger.info(
+            {
+                "session Id": data.session_id,
+                "song Id": data.song_id,
+                "thumbs up": data.thumbs_up,
+                "thumbs down": data.thumbs_down,
+            }
+        )
