+++ back/src/utils.py
@@ -1,6 +1,4 @@
 import os
-import ast
-import pandas as pd
 
 
 def create_dir(path: str) -> None:
@@ -9,19 +7,3 @@ def create_dir(path: str) -> None:
             os.makedirs(path)
     except OSError:
         raise Exception(f"Creating {path} is failed !!!")
-
-
-def get_first_dir(path: str) -> str:
-    dirs = os.listdir(path)
-    return dirs[0]
-
-
-def str2list(data: pd.DataFrame, columns: list) -> None:
-    for col in columns:
-        data[col] = data[col].apply(lambda x: ast.literal_eval(x))
-
-
-def check_substring(result: str, query: str) -> bool:
-    result = result.lower().replace(" ", "")
-    query = query.lower().replace(" ", "")
-    return result in query or query in result
